requirements.txt
flask==2.2.2
flask_login==0.6.2
Markdown==3.4.1
passlib==1.7.4
requests==2.28.1
pycryptodome==3.16.0
bleach==5.0.1
Jinja2==3.0.3
Flask-Limiter==3.1.0
python-memcached==1.59
password-strength==0.0.3.post2
Flask-WTF==1.1.1
uwsgi==2.0.21